Switch to a dedicated class to represent all possible options.
This is more generalizable and allows us to easily add more options
without bloating the constructors. For example, we can add a
'require_minimum_cache' option to ensure that the cache is always large
enough for efficient (dense) extraction.

We also add an overload for each constructor that does not require the
options object, for simplicity.