Ensure thread-safe destruction of HDF5 members in the Extractor.
Apparently that needs to be protected by the lock as well, otherwise we
get cryptic HDF5 errors after the first use of the HDF5 library. This
also implies that we need to store all HDF5 members in a pointer so that
we can manually control their destruction in the Extractor's destructor.