Improved primary sparse extraction by more carefully constructing hyperslabs. (#19)
For block and indexed extractions, we extract the indices first and use them to
identify the subset of values that we need. This avoids unnecessarily
extracting the full set of values for each primary dimension element.

Also did some minor refactoring of the primary extraction tests to better
enable re-use of the HDF5 file creation code.