Streamlined the sparse test suite.
We split off the uncompressed sparse tests into its own suite; this avoids
reepating the parametrized tests for something that doesn't really affect us
and should be handled by the HDF5 library.

We remove the 0.001 cache_fraction option; this is redundant with the 0 option
now that all sparse extractors implicitly require a minimum cache (and would
not respect any require_minimum_cache=false setting).

We also check multiple cache_fraction options in the cache type test suite, for
consistency with the other tests; this requires some work to turn down the
matrix size to offset the slowness of secondary extraction.