More careful chunk caching for sparse matrices.
This aims to keep at least two chunks: one overlapping the start of the range
of the primary dimension element, and another overlapping the end of the range,
so that iteration in any direction from the current primary dimension element
will not have to re-read the same chunk from file.

To pass the chunk cache size down to the bottom-level extractor class, we move
the cache parameters into the MatrixDetails class. Happily enough, this also
simplifies a lot of the constructor code.