Renamed classes and functions for more consistency, less redundancy. (#21)
This removes the "hdf5" component in the class and function names. It should be
obvious that we're dealing with HDF5 given that we're in the tatami_hdf5
namespace, so I don't see the need to repeat it. Also clarified that the sparse
readers/writers are operating on compressed sparse matrices.

The various *Options and *Parameters classes now follow the format of
"<class/function name>Options". We use different Options classes for dense and
sparse matrix constructors to future-proof for class-specific parameters. 

The nested enums have been moved out of the WriteSparseMatrixParameters class,
mostly for easier writing by callers (the fully namespaced name is pretty long)
but also to potentially allow re-use in a future dense matrix writer.

Also removed the Stored class in favor of tatami::ElementType.