Use the SUBPAR_USES_OPENMP macro when choosing the lock method.
This is motivated by tatami's switch to using the subpar library for
its parallelization, which auto-chooses between OpenMP and <thread>.
By checking the macro, we make sure that we use a locking scheme that is
appropriate to the parallelization scheme chosen by tatami::parallelize,
a.k.a., the new subpar::parallelize. (Unless overridden by macros, in
which case all bets are off and the developer should figure it out.)

Also avoid include's of the standard headers inside our own namespace.