Directly fill the cache in a single HDF5 call in dense oracular extractors.
This refactors the cache to become a shared memory pool. During a populate()
cycle for the next set of future slabs, the re-used slabs are defragmented to
make a contiguous allocation of available cache memory. This contiguous
allocation is filled with the contents of the to-be-populated slabs via a
single call to the HDF5 library with hyperslab unions corresponding to those
slabs. The aim is to reduce the overhead from separate calls to the HDF5
library, equivalent to the primary sparse extractors.

Note that this optimization is only available when the target dimension
corresponds to HDF5 rows. If transposition is required, we do separate calls
for each slab, we can't afford to hold all slabs in a separate buffer before
performing the transposition into the cache pool. To keep these code paths
distinct, we split the oracular DenseMatrix extractors into two subclasses
depending on whether the extraction requires transposition. We do the same to
all other DenseMatrix extractors for consistency.