Use dataset-specific chunk caches for sparse primary extractors. (#26)
This avoids allocating a larger-than-necessary cache, e.g., if the index
dataset has smaller data elements than the value dataset.

We increase the minimum number of chunks to 3. For the common case of a linear
scan across primary dimension elements, this bump gives us a chance of
retaining partially-read chunks if the library loads them in random order. 

We also set a minimum cache size of 1 MB so that we are at least as good as the
default settings when the access pattern is not linear.