Do not use hyperslab unions as they are too slow. (#27)
As of 1.14.5, they are still painfully slow. Instead, just do multiple calls to
the HDF5 library when dealing with non-contiguous spaces. For indexed extraction,
we read the entire bounding interval and then extract the indices of interest.

Also explain why a pretend chunk length of 1 isn't so easy; removed unnecessary
checks for empty refined range; and get some test coverage for empty indices.